auth commit

+++ laravel/app/Http/Controllers/ProjectController.php
@@ -6,7 +6,7 @@ use Illuminate\Http\Request;
 use App\Models\Project;
 use App\Models\SustainabilityCategory;
 use App\Models\SustainabilityMetric;
-
+use Illuminate\Support\Facades\Auth;
 
 class ProjectController extends Controller
 {
@@ -24,6 +24,9 @@ class ProjectController extends Controller
      */
     public function create()
     {
+        if (Auth::user()->accountType == 'General User' or Auth::user()->accountType == 'Analyst'){
+            abort(403);
+        }
         return view('projects/create');
     }
 
@@ -32,10 +35,11 @@ class ProjectController extends Controller
      */
     public function store(Request $request)
     {
-
+       
         $project = new Project();
         $project->project_name = $request->project_name;
         $project->description = $request->description;
+        $project->budget = $request->budget;
         $project->start_date = $request->start_date;
         $project->end_date = $request->end_date;
         $project->overall_sustainability_score = $request->overall_sustainability_score;
@@ -58,6 +62,13 @@ class ProjectController extends Controller
      */
     public function edit(string $id)
     {
+        if (Auth::guest()){
+            return redirect('/users/login');
+        }
+        if (Auth::user()->accountType == 'General User' or Auth::user()->accountType == 'Analyst'){
+            abort(403);
+        }
+
         $project = Project::find($id);
         return view('projects/edit', ['project' => $project]);
     }
@@ -71,6 +82,7 @@ class ProjectController extends Controller
         $project = Project::find($request->id);
         $project->project_name = $request->project_name;
         $project->description = $request->description;
+        $project->budget = $request->budget;
         $project->start_date = $request->start_date;
         $project->end_date = $request->end_date;
         $project->overall_sustainability_score = $request->overall_sustainability_score;
@@ -81,8 +93,20 @@ class ProjectController extends Controller
 
     public function destroy(string $id)
     {
+        if (Auth::user()->accountType == 'General User' or Auth::user()->accountType == 'Analyst'){
+            abort(403);
+        }
         $project = Project::find($id);
         $project->delete();
         return redirect('/projects');
+    }
+    public function prediction(string $id)
+    {   
+        if (Auth::user()->accountType == 'General User'){
+            abort(403);
+        }
+        dd('prediction');
+        $project = Project::find($id);
+        return view('projects/show', ['project' => $project]);
     }
 }
